update integration sandbox
