--- examples/minimal/packages/client-react/src/mud/clientComponents.ts
@@ -1,1 +1,0 @@
-export const clientComponents = {};
