--- examples/minimal/packages/client-react/src/mud/supportedChains/types.ts
@@ -1,3 +1,0 @@
-import { Chain } from "@wagmi/chains";
-
-export type MudChain = Chain & { modeUrl?: string; faucetUrl?: string };
