+++ integration-sandbox/packages/client-vanilla/package.json
@@ -12,7 +12,7 @@
   "dependencies": {
     "@ethersproject/providers": "^5.7.2",
     "@improbable-eng/grpc-web": "^0.15.0",
-    "@latticexyz/cli": "link:../../../packages/cli",
+    "@latticexyz/common": "link:../../../packages/common",
     "@latticexyz/dev-tools": "link:../../../packages/dev-tools",
     "@latticexyz/network": "link:../../../packages/network",
     "@latticexyz/recs": "link:../../../packages/recs",
