--- integration-sandbox/packages/client-vanilla/src/mud/clientComponents.ts
@@ -1,1 +1,0 @@
-export const clientComponents = {};
