+++ integration-sandbox/packages/client-vanilla/src/mud/createClientComponents.ts
@@ -1,0 +1,10 @@
+import { SetupNetworkResult } from "./setupNetwork";
+
+export type ClientComponents = ReturnType<typeof createClientComponents>;
+
+export function createClientComponents({ components }: SetupNetworkResult) {
+  return {
+    ...components,
+    // add your client components or overrides here
+  };
+}
