+++ integration-sandbox/packages/client-vanilla/src/mud/getNetworkConfig.ts
@@ -1,20 +1,18 @@
 import { SetupContractConfig, getBurnerWallet } from "@latticexyz/std-client";
-import { foundry } from "@wagmi/chains";
-import latticeTestnet from "./supportedChains/latticeTestnet";
-import { MudChain } from "./supportedChains/types";
 import worldsJson from "contracts/worlds.json";
+import { supportedChains } from "./supportedChains";
 
 const worlds = worldsJson as Partial<Record<string, { address: string; blockNumber?: number }>>;
 
 type NetworkConfig = SetupContractConfig & {
   privateKey: string;
   faucetServiceUrl?: string;
+  snapSync?: boolean;
 };
 
 export async function getNetworkConfig(): Promise<NetworkConfig> {
   const params = new URLSearchParams(window.location.search);
 
-  const supportedChains: MudChain[] = [foundry, latticeTestnet];
   const chainId = Number(params.get("chainId") || import.meta.env.VITE_CHAIN_ID || 31337);
   const chainIndex = supportedChains.findIndex((c) => c.id === chainId);
   const chain = supportedChains[chainIndex];
@@ -49,6 +47,7 @@ export async function getNetworkConfig(): Promise<NetworkConfig> {
     faucetServiceUrl: params.get("faucet") ?? chain.faucetUrl,
     worldAddress,
     initialBlockNumber,
+    snapSync: params.get("snapSync") === "true",
     disableCache: params.get("cache") === "false",
   };
 }
