+++ integration-sandbox/packages/client-vanilla/src/mud/setup.ts
@@ -1,94 +1,16 @@
-import { setupMUDV2Network } from "@latticexyz/std-client";
-import { createFastTxExecutor, createFaucetService } from "@latticexyz/network";
-import { getNetworkConfig } from "./getNetworkConfig";
-import { defineContractComponents } from "./contractComponents";
-import { clientComponents } from "./clientComponents";
-import { world } from "./world";
-import { Contract, Signer, utils } from "ethers";
-import { JsonRpcProvider } from "@ethersproject/providers";
-import { IWorld__factory } from "contracts/types/ethers-contracts/factories/IWorld__factory";
-import storeConfig from "contracts/mud.config";
+import { createClientComponents } from "./createClientComponents";
+import { createSystemCalls } from "./createSystemCalls";
+import { setupNetwork } from "./setupNetwork";
 
 export type SetupResult = Awaited<ReturnType<typeof setup>>;
 
 export async function setup() {
-  const contractComponents = defineContractComponents(world);
-  const networkConfig = await getNetworkConfig();
-  const result = await setupMUDV2Network<typeof contractComponents, typeof storeConfig>({
-    networkConfig,
-    world,
-    contractComponents,
-    syncThread: "main",
-    storeConfig,
-    worldAbi: IWorld__factory.abi,
-  });
-
-  result.startSync();
-
-  // Request drip from faucet
-  const signer = result.network.signer.get();
-  if (networkConfig.faucetServiceUrl && signer) {
-    const address = await signer.getAddress();
-    console.info("[Dev Faucet]: Player address -> ", address);
-
-    const faucet = createFaucetService(networkConfig.faucetServiceUrl);
-
-    const requestDrip = async () => {
-      const balance = await signer.getBalance();
-      console.info(`[Dev Faucet]: Player balance -> ${balance}`);
-      const lowBalance = balance?.lte(utils.parseEther("1"));
-      if (lowBalance) {
-        console.info("[Dev Faucet]: Balance is low, dripping funds to player");
-        // Double drip
-        await faucet.dripDev({ address });
-        await faucet.dripDev({ address });
-      }
-    };
-
-    requestDrip();
-    // Request a drip every 20 seconds
-    setInterval(requestDrip, 20000);
-  }
-
-  // Create a World contract instance
-  const worldContract = IWorld__factory.connect(
-    networkConfig.worldAddress,
-    signer ?? result.network.providers.get().json
-  );
-
-  // Create a fast tx executor
-  const fastTxExecutor =
-    signer?.provider instanceof JsonRpcProvider
-      ? await createFastTxExecutor(signer as Signer & { provider: JsonRpcProvider })
-      : null;
-
-  // TODO: infer this from fastTxExecute signature?
-  type BoundFastTxExecuteFn<C extends Contract> = <F extends keyof C>(
-    func: F,
-    args: Parameters<C[F]>,
-    options?: {
-      retryCount?: number;
-    }
-  ) => Promise<ReturnType<C[F]>>;
-
-  function bindFastTxExecute<C extends Contract>(contract: C): BoundFastTxExecuteFn<C> {
-    return async function (...args) {
-      if (!fastTxExecutor) {
-        throw new Error("no signer");
-      }
-      const { tx } = await fastTxExecutor.fastTxExecute(contract, ...args);
-      return await tx;
-    };
-  }
-
+  const network = await setupNetwork();
+  const components = createClientComponents(network);
+  const systemCalls = createSystemCalls(network, components);
   return {
-    ...result,
-    components: {
-      ...result.components,
-      ...clientComponents,
-    },
-    worldContract,
-    worldSend: bindFastTxExecute(worldContract),
-    fastTxExecutor,
+    network,
+    components,
+    systemCalls,
   };
 }
