+++ integration-sandbox/packages/client-vanilla/src/mud/supportedChains.ts
@@ -1,0 +1,5 @@
+import { MUDChain, latticeTestnet } from "@latticexyz/common/chains";
+import { foundry } from "@wagmi/chains";
+
+// If you are deploying to chains other than anvil or Lattice testnet, add them here
+export const supportedChains: MUDChain[] = [foundry, latticeTestnet];
